Streamlit_Gemini_Health_Diagnostics/requirements.txt
streamlit
google-generativeai
pdfplumber
pytesseract
Pillow
python-dotenv
